+++ src/scripts/InstanceScripts/Instance_SethekkHalls.cpp
@@ -1995,7 +1995,7 @@ class TalonKingIkissAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
@@ -2038,7 +2038,7 @@ class TalonKingIkissAI : public CreatureAIScript
                 return;
             }
 
-            auto random_index = RandomUInt(0, TargetTable.size() - 1);
+            auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
             auto random_target = TargetTable[random_index];
 
             if (random_target == nullptr)
