+++ src/scripts/InstanceScripts/Raid_BlackTemple.cpp
@@ -1688,7 +1688,7 @@ class NajentusAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
@@ -2201,7 +2201,7 @@ class GurtoggAI : public CreatureAIScript
                     if (!TargetTable.size())
                         return;
 
-                    auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                    auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                     auto random_target = TargetTable[random_index];
 
                     if (random_target == nullptr)
@@ -2321,7 +2321,7 @@ class GurtoggAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
@@ -3002,7 +3002,7 @@ class ShahrazAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
@@ -3422,7 +3422,7 @@ class TeronGorefiendAI : public CreatureAIScript
                 if (!TargetTable.size())
                     return;
 
-                auto random_index = RandomUInt(0, TargetTable.size() - 1);
+                auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
                 auto random_target = TargetTable[random_index];
 
                 if (random_target == nullptr)
