+++ src/scripts/InstanceScripts/Raid_IceCrownCitadel.cpp
@@ -351,7 +351,7 @@ class LordMarrowgarAI : public MoonScriptBossAI
             if (!TargetTable.size())
                 return;
 
-            auto random_index = RandomUInt(0, TargetTable.size() - 1);
+            auto random_index = RandomUInt(0, uint32(TargetTable.size() - 1));
             auto random_target = TargetTable[random_index];
 
             if (random_target == nullptr)
